+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Support database as schema source.
+
 ## 0.2.0
 - Support text schemas.
 - Output the diagram as PNG.